refactor: moved AddressModal to components folder
