+++ src/components/AddressModal/AddressModal.jsx
@@ -1,8 +1,8 @@
 import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import { toast } from "react-toastify";
-import { useAuth } from "../../../../context";
-import { addNewAddressHandler, updateAddressHandler } from "../../../../utils";
+import { useAuth } from "../../context";
+import { addNewAddressHandler, updateAddressHandler } from "../../utils";
 import "./AddressModal.css";
 
 const AddressModal = ({ editAddress, setEditAddress, showAddressModal, setShowAddressModal }) => {
